Allow disabling supported extensions at runtime.
Extension support is gated behind trait `const bool`s to save code size when some are not required. However, some execution environments might want to allow some extensions to be disabled and re-enabled at runtime. This commit adds `fn(&self) -> bool` functions to the `ExecutionEnvironment` trait that allow this. The default implementation just trivially returns `true` for all extensions all the time, which means the check is optimized out of the code. Thus, there is no impact on `ExecutionEnvironment`s that don't use this feature.